feat(useVModel): use call function
