+++ packages/varlet-ui/src/fab/Fab.tsx
@@ -1,8 +1,8 @@
 import Button from '../button'
 import Icon from '../icon'
 import { defineComponent, Ref, ref, Teleport, Transition, watch } from 'vue'
-import { useClickOutside, useVModel } from '@varlet/use'
-import { call, createNamespace, flatFragment, useTeleport } from '../utils/components'
+import { useClickOutside } from '@varlet/use'
+import { call, createNamespace, flatFragment, useTeleport, useVModel } from '../utils/components'
 import { toSizeUnit } from '../utils/elements'
 import { toNumber } from '@varlet/shared'
 import { props } from './props'
