+++ packages/varlet-ui/src/menu/usePopover.ts
@@ -1,8 +1,8 @@
 import flip from '@popperjs/core/lib/modifiers/flip'
 import offset from '@popperjs/core/lib/modifiers/offset'
-import { useClickOutside, useVModel } from '@varlet/use'
+import { useClickOutside } from '@varlet/use'
 import { doubleRaf, toPxNum } from '../utils/elements'
-import { call } from '../utils/components'
+import { call, useVModel } from '../utils/components'
 import { onMounted, onUnmounted, ref, watch, type Ref } from 'vue'
 import { createPopper } from '@popperjs/core/lib/popper-lite'
 import { useZIndex } from '../context/zIndex'
