+++ packages/varlet-use/src/index.ts
@@ -3,4 +3,4 @@ export * from './useClickOutside.js'
 export * from './useMounted.js'
 export * from './useParent.js'
 export * from './useChildren.js'
-export * from './useVModel.js'
+
